ad_observatory_api_backend/requirements.txt
appdirs==1.4.4
Authlib==0.15.2
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
dhash==1.3
distlib==0.3.1
distro==1.5.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
gunicorn==20.0.4
html5lib==1.1
humanize==3.1.0
idna==2.10
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
memoization==0.3.1
msgpack==1.0.0
numpy==1.19.4
packaging==20.7
pandas==1.1.5
pep517==0.9.1
Pillow==8.0.1
progress==1.5
psycopg2-binary==2.8.6
pybktree==1.1
pycountry==20.7.3
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.4
requests==2.25.0
retrying==1.3.3
simplejson==3.17.2
six==1.15.0
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
